+++ packages/plugin-hardhat/src/prepare-upgrade.ts
@@ -1,7 +1,7 @@
 import { HardhatRuntimeEnvironment } from 'hardhat/types';
 import type { ContractFactory } from 'ethers';
 
-import { ValidationOptions, withValidationDefaults, setProxyKind } from '@openzeppelin/upgrades-core';
+import { ValidationOptions } from '@openzeppelin/upgrades-core';
 
 import { ContractAddressOrInstance, deployImpl, getContractAddress } from './utils';
 
@@ -13,12 +13,8 @@ export type PrepareUpgradeFunction = (
 
 export function makePrepareUpgrade(hre: HardhatRuntimeEnvironment): PrepareUpgradeFunction {
   return async function prepareUpgrade(proxy, ImplFactory, opts: ValidationOptions = {}) {
-    const { provider } = hre.network;
-
     const proxyAddress = getContractAddress(proxy);
-
-    await setProxyKind(provider, proxyAddress, opts);
-
-    return await deployImpl(hre, ImplFactory, withValidationDefaults(opts), proxyAddress);
+    const { impl } = await deployImpl(hre, ImplFactory, opts, proxyAddress);
+    return impl;
   };
 }