+++ packages/plugin-truffle/src/utils/deploy-impl.ts
@@ -7,40 +7,52 @@ import {
   Manifest,
   getImplementationAddress,
   getStorageLayoutForAddress,
+  inferProxyKind,
+  ValidationOptions,
+  setProxyKind,
 } from '@openzeppelin/upgrades-core';
 
 import { deploy } from './deploy';
-import { Options } from './options';
+import { Options, withDefaults } from './options';
 import { ContractClass, getTruffleConfig } from './truffle';
 import { validateArtifacts, getLinkedBytecode } from './validations';
 import { wrapProvider } from './wrap-provider';
 
-export async function deployImpl(
-  Contract: ContractClass,
-  requiredOpts: Required<Options>,
-  proxyAddress?: string,
-): Promise<string> {
-  if (requiredOpts.kind === 'transparent') {
-    requiredOpts.unsafeAllow.push('missing-public-upgradeto');
-  }
+interface DeployedImpl {
+  impl: string;
+  kind: NonNullable<ValidationOptions['kind']>;
+}
 
-  const provider = wrapProvider(requiredOpts.deployer.provider);
+export async function deployImpl(Contract: ContractClass, opts: Options, proxyAddress?: string): Promise<DeployedImpl> {
+  const fullOpts = withDefaults(opts);
+  const provider = wrapProvider(fullOpts.deployer.provider);
   const { contracts_build_directory, contracts_directory } = getTruffleConfig();
   const validations = await validateArtifacts(contracts_build_directory, contracts_directory);
   const linkedBytecode = await getLinkedBytecode(Contract, provider);
   const version = getVersion(Contract.bytecode, linkedBytecode);
   const layout = getStorageLayout([validations], version);
-  assertUpgradeSafe([validations], version, requiredOpts);
+
+  if (opts.kind === undefined) {
+    fullOpts.kind = inferProxyKind(validations, version);
+  }
+
+  if (proxyAddress !== undefined) {
+    await setProxyKind(provider, proxyAddress, fullOpts);
+  }
+
+  assertUpgradeSafe([validations], version, fullOpts);
 
   if (proxyAddress !== undefined) {
     const manifest = await Manifest.forNetwork(provider);
     const currentImplAddress = await getImplementationAddress(provider, proxyAddress);
     const currentLayout = await getStorageLayoutForAddress(manifest, validations, currentImplAddress);
-    assertStorageUpgradeSafe(currentLayout, layout, requiredOpts);
+    assertStorageUpgradeSafe(currentLayout, layout, fullOpts);
   }
 
-  return await fetchOrDeploy(version, provider, async () => {
-    const deployment = await deploy(requiredOpts.deployer, Contract);
+  const impl = await fetchOrDeploy(version, provider, async () => {
+    const deployment = await deploy(fullOpts.deployer, Contract);
     return { ...deployment, layout };
   });
+
+  return { impl, kind: fullOpts.kind };
 }