+++ packages/plugin-hardhat/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Add option `unsafeAllowRenames` to configure storage layout check to allow variable renaming.
+
 ## 1.8.2 (2021-05-13)
 
 - Simplify `deployTransaction` assignment in `deployProxy` as a fix for congested networks.